feat: add read limiter

+++ src/app/api/options/route.ts
@@ -4,10 +4,17 @@ import { authenticate } from '../utils/decode';
 import { Category } from '../models/Category';
 import { CategoryOption } from '../models/Category_Options';
 import { category_option_schema } from '../validators/validate';
+import { rateLimitMiddleware } from '../utils/limiter';
 
 export async function POST(request: any) {
-  // let flag = await authenticate(request)
-  // if (!flag) return NextResponse.json({ status: 'Error', message: 'Access Denied. Invalid Token.' }, { status: 400 });
+  let pass = await rateLimitMiddleware(request);
+  if (!pass)
+    return NextResponse.json(
+      { status: 'Error', message: 'Too Many Requests.' },
+      { status: 400 }
+    );
+  let flag = await authenticate(request)
+  if (!flag) return NextResponse.json({ status: 'Error', message: 'Access Denied. Invalid Token.' }, { status: 400 });
   const {error, value} = await category_option_schema.validate(await request.json())
   if(error) return NextResponse.json({ message: error.details[0].message }, { status: 400 });
   let options: any =  value;
@@ -17,16 +24,28 @@ export async function POST(request: any) {
 }
 
 export async function GET(request: any) {
-  // let flag = await authenticate(request)
-  // if (!flag) return NextResponse.json({ status: 'Error', message: 'Access Denied. Invalid Token.' }, { status: 400 });
+  let pass = await rateLimitMiddleware(request);
+  if (!pass)
+    return NextResponse.json(
+      { status: 'Error', message: 'Too Many Requests.' },
+      { status: 400 }
+    );
+  let flag = await authenticate(request)
+  if (!flag) return NextResponse.json({ status: 'Error', message: 'Access Denied. Invalid Token.' }, { status: 400 });
   await dbConnect();
   let options: any[] = await CategoryOption.find().populate('category');
   return NextResponse.json({ options });
 }
 
 export async function DELETE(request: any) {
-  // let flag = await authenticate(request)
-  // if (!flag) return NextResponse.json({ status: 'Error', message: 'Access Denied. Invalid Token.' }, { status: 400 });
+  let pass = await rateLimitMiddleware(request);
+  if (!pass)
+    return NextResponse.json(
+      { status: 'Error', message: 'Too Many Requests.' },
+      { status: 400 }
+    );
+  let flag = await authenticate(request)
+  if (!flag) return NextResponse.json({ status: 'Error', message: 'Access Denied. Invalid Token.' }, { status: 400 });
   const id = request.nextUrl.searchParams.get('id');
   await dbConnect();
   await CategoryOption.findByIdAndDelete(id);
